requirements.txt
streamlit
google-generativeai
SpeechRecognition
gTTS
streamlit-mic-recorder
pydub
pandas